# api/swagger.yaml
openapi: 3.0.0
info:
  title: API Artículo 140 - Sistema VOAE
  version: 1.0.0
  description: API REST para la gestión del Sistema de Gestión de Horas VOAE (Vinculación con la Sociedad) de la UNAH
  contact:
    name: MelvinMajano
    url: https://github.com/MelvinMajano/api-articulo140
  license:
    name: ISC
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único del usuario
        name:
          type: string
          description: Nombre completo del usuario
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        accountNumber:
          type: integer
          description: Número de cuenta del estudiante
        identityNumber:
          type: string
          description: Número de identidad
        role:
          type: string
          enum: [admin, student, supervisor]
          description: Rol del usuario
        degreeId:
          type: integer
          description: ID de la carrera asociada
        isDeleted:
          type: string
          enum: [true, false]
          description: Estado de eliminación
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de error
        error:
          type: string
          description: Detalles del error
        status:
          type: integer
          description: Código de estado HTTP
    Success:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de éxito
        data:
          type: object
          description: Datos de respuesta
security:
  - bearerAuth: []
paths:
...
  /api/activities/{activityid}/register/{id}:
    post:
      tags:
        - Inscripciones
      summary: Registrar estudiante en actividad
      description: Inscribir un estudiante en una actividad específica
      parameters:
        - name: activityid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la actividad
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del estudiante
      responses:
        '201':
          description: Estudiante registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Error en el registro (cupos llenos, ya registrado, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acceso denegado - Solo estudiantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Actividad o estudiante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/activities/{activityid}/register:
    get:
      tags:
        - Inscripciones
      summary: Obtener estudiantes registrados en actividad
      description: Listar todos los estudiantes inscritos en una actividad específica
      parameters:
        - name: activityid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la actividad
      responses:
        '200':
          description: Lista de estudiantes registrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  students:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/User'
                        - type: object
                          properties:
                            registrationDate:
                              type: string
                              format: date-time
        '403':
          description: Acceso denegado - Solo supervisores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Actividad no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/activities/register/end/{id}:
    put:
      tags:
        - Inscripciones
      summary: Cerrar inscripciones
      description: Cerrar el periodo de inscripciones para una actividad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la actividad
      responses:
        '200':
          description: Inscripciones cerradas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '403':
          description: Acceso denegado - Solo supervisores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Actividad no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/activities/finish/{id}:
    put:
      tags:
        - Inscripciones
      summary: Finalizar actividad
      description: Marcar una actividad como finalizada
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la actividad
      responses:
        '200':
          description: Actividad finalizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '403':
          description: Acceso denegado - Solo supervisores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Actividad no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ATTENDANCE ENDPOINTS
...